Sligthly more share stuff
